pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datalex-cli"
version = "1.4.0"
description = "Git-native data modeling for dbt users"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "DuckCode AI Labs", email = "hello@duckcode.ai" }]
keywords = ["dbt", "data-modeling", "yaml", "git-native", "erd", "data-warehouse"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Database",
  "Topic :: Software Development :: Code Generators",
]
dependencies = [
  "PyYAML>=6.0",
  "jsonschema>=4.0",
]

[project.optional-dependencies]
# Bundled Node runtime — `pip install datalex-cli[serve]` gives the user
# a working `datalex serve` even on hosts without Node on PATH.
serve = ["nodejs-wheel>=20,<23"]
duckdb = ["duckdb>=0.9"]
postgres = ["psycopg2-binary"]
mysql = ["mysql-connector-python"]
snowflake = ["snowflake-connector-python"]
bigquery = ["google-cloud-bigquery"]
databricks = ["databricks-sql-connector"]
sqlserver = ["pyodbc"]
redshift = ["redshift-connector"]
all = [
  "nodejs-wheel>=20,<23",
  "duckdb>=0.9",
  "psycopg2-binary",
  "mysql-connector-python",
  "snowflake-connector-python",
  "google-cloud-bigquery",
  "databricks-sql-connector",
  "pyodbc",
  "redshift-connector",
]

[project.urls]
Homepage = "https://duckcode.ai"
Repository = "https://github.com/duckcode-ai/DataLex"
Issues = "https://github.com/duckcode-ai/DataLex/issues"

[project.scripts]
datalex = "datalex_cli.main:main"
dm = "datalex_cli.main:main"

[tool.setuptools.packages.find]
where = [
  "packages/core_engine/src",
  "packages/cli/src",
  "packages/readiness_engine/src",
]

[tool.setuptools.package-data]
# `_server` and `_webapp` ship the Node api-server entry point and the
# built web-app dist respectively. Both are populated by the release
# workflow before `python -m build`, so source checkouts without a web
# build still install — `datalex serve` then falls back to the
# repo-local paths.
"datalex_core" = [
  "_schemas/**/*.json",
  "_server/**/*",
  "_webapp/**/*",
]
